Separate recipe edit page from view page, save ingredients on blur
- Create /recipes/[id]/edit/page.tsx as a dedicated edit page (always in
  edit mode, no toggle). Navigating to /edit from the view page or after
  recipe creation.
- Simplify /recipes/[id]/page.tsx to view-only (no isEditMode state,
  no inline editing). Edit recipe menu item navigates to /edit.
- Ingredients now save on blur instead of requiring a checkmark click.
  The save button is removed from SortableIngredientItem.
- Fix ingredient order: useUpdateRecipeIngredient now uses optimistic
  updates (onMutate setQueryData) so the list position is preserved
  immediately without a server refetch reordering items.
- Update /recipes/new/page.tsx to redirect to /recipes/[id]/edit instead
  of /recipes/[id]?edit=true after creation or URL import.
- Update and split tests: page.test.tsx covers view-only behavior,
  edit/page.test.tsx covers all edit functionality.
